# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Groupe Acharaf luxury real estate website
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects (optionally filter by brand)
      parameters:
        - name: brandId
          in: query
          required: false
          schema:
            type: integer
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [upcoming, ongoing, completed]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Articles ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Brand:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        tagline:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        segment:
          type: string
          enum: [luxury, premium, social]
        logoUrl:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        accentColor:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - segment
        - createdAt
...
    Project:
      type: object
      properties:
        id:
          type: integer
        brandId:
          type: integer
        brand:
          $ref: "#/components/schemas/Brand"
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        location:
          type: string
        city:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        priceMin:
          type: number
        priceMax:
          type: number
        surfaceMin:
          type: number
        surfaceMax:
          type: number
        deliveryDate:
          type: string
        featured:
          type: boolean
        coverImageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - brandId
        - title
        - slug
        - status
        - featured
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        brandId:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        location:
          type: string
        city:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        priceMin:
          type: number
        priceMax:
          type: number
        surfaceMin:
          type: number
        surfaceMax:
          type: number
        deliveryDate:
          type: string
        featured:
          type: boolean
        coverImageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
      required:
        - brandId
        - title
        - slug
        - status
...